+++ src/imageUtils/index.ts
@@ -1,0 +1,2 @@
+export { mergeImages } from './mergeImages'
+export { writeImages } from './writeImages'
