+++ src/imageUtils/mergeImages.ts
@@ -5,9 +5,15 @@ type ImgData = Readonly<{
   y: number
 }>
 
-export const mergeImages = (imgs: ReadonlyArray<Jimp>): Jimp => {
+/**
+ * Merges an array of Jimp images into a single image.
+ *
+ * @param images - An array of Jimp images to be merged.
+ * @returns A Jimp image that is the result of merging all input images.
+ */
+export function mergeImages(images: ReadonlyArray<Jimp>): Jimp {
   let imgHeight = 0
-  const imgData: ImgData[] = imgs.map((img) => {
+  const imgData: ImgData[] = images.map((img) => {
     const res = { img, y: imgHeight }
     imgHeight += img.bitmap.height
     return res
