+++ src/index.ts
@@ -6,3 +6,4 @@ export {
   CompareOptions,
   HighlightColor,
 } from './compare-pdf-to-snapshot'
+export { Pdf2PngOpts } from './types'
