--- src/pdf2png.ts
@@ -1,159 +1,0 @@
-import * as Canvas from 'canvas'
-import * as assert from 'assert'
-import * as fs from 'fs'
-import * as path from 'path'
-import * as pdfjsLib from 'pdfjs-dist/legacy/build/pdf'
-import Jimp, { read } from 'jimp'
-import { mergeImages } from './merge-images'
-
-function convertFromMmToPx(sizeMm: number, dpi: number): number {
-  if (sizeMm <= 0 || dpi <= 0) {
-    return 0
-  }
-  const sizeInch = sizeMm / 25.4
-  return Math.round(sizeInch * dpi)
-}
-function convertFromPxToMm(sizePx: number, dpi: number): number {
-  if (sizePx <= 0 || dpi <= 0) {
-    return 0
-  }
-  const sizeInch = sizePx / dpi
-  return Math.round(sizeInch * 25.4)
-}
-
-type CanvasAndContext = {
-  canvas: Canvas.Canvas
-  context: Canvas.CanvasRenderingContext2D
-}
-
-class NodeCanvasFactory {
-  create(width: number, height: number): CanvasAndContext {
-    assert.ok(width > 0 && height > 0, 'Invalid canvas size')
-    const canvas = Canvas.createCanvas(width, height)
-    const context = canvas.getContext('2d')
-    return {
-      canvas,
-      context,
-    }
-  }
-
-  reset(canvasAndContext: CanvasAndContext, width: number, height: number): void {
-    assert.ok(canvasAndContext.canvas, 'Canvas is not specified')
-    assert.ok(width > 0 && height > 0, 'Invalid canvas size')
-    canvasAndContext.canvas.width = width
-    canvasAndContext.canvas.height = height
-  }
-
-  destroy(canvasAndContext: CanvasAndContext): void {
-    assert.ok(canvasAndContext.canvas, 'Canvas is not specified')
-    canvasAndContext.canvas.width = 0
-    canvasAndContext.canvas.height = 0
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    canvasAndContext.canvas = null
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    canvasAndContext.context = null
-  }
-}
-
-// pdfjs location
-const PDFJS_DIR = path.dirname(require.resolve('pdfjs-dist'))
-
-// Some PDFs need external cmaps.
-const CMAP_URL = path.join(PDFJS_DIR, '../cmaps/')
-const CMAP_PACKED = true
-
-// Where the standard fonts are located.
-const STANDARD_FONT_DATA_URL = path.join(PDFJS_DIR, '../standard_fonts/')
-
-/** The options type for {@link pdf2png}. */
-export type Pdf2PngOpts = {
-  /**
-   * Indicates whether to upscale the image resolution from 72 DPI to 144 DPI.
-   * Enabling this option results in a slower conversion process but produces a higher resolution image.
-   *
-   * @defaultValue true
-   */
-  scaleImage?: boolean
-}
-
-const pdf2PngDefOpts: Required<Pdf2PngOpts> = {
-  scaleImage: true,
-}
-
-function getPageViewPort(page: pdfjsLib.PDFPageProxy, scaleImage: boolean): pdfjsLib.PageViewport {
-  const viewport = page.getViewport({ scale: 1.0 })
-  if (scaleImage === false) {
-    return viewport
-  }
-
-  // Increase resolution
-  const horizontalMm = convertFromPxToMm(viewport.width, 72)
-  const verticalMm = convertFromPxToMm(viewport.height, 72)
-  const actualWidth = convertFromMmToPx(horizontalMm, 144)
-  const actualHeight = convertFromMmToPx(verticalMm, 144)
-  const scale = Math.min(actualWidth / viewport.width, actualHeight / viewport.height)
-  return page.getViewport({ scale })
-}
-
-export async function pdf2png(
-  pdf: string | Buffer,
-  options: Pdf2PngOpts = {},
-): Promise<ReadonlyArray<Jimp>> {
-  const opts = {
-    ...pdf2PngDefOpts,
-    ...options,
-  }
-
-  // Load PDF
-  const data = new Uint8Array(Buffer.isBuffer(pdf) ? pdf : fs.readFileSync(pdf))
-  const loadingTask = pdfjsLib.getDocument({
-    data,
-    cMapUrl: CMAP_URL,
-    cMapPacked: CMAP_PACKED,
-    standardFontDataUrl: STANDARD_FONT_DATA_URL,
-  })
-
-  const pdfDocument = await loadingTask.promise
-  const numPages = pdfDocument.numPages
-
-  const canvasFactory = new NodeCanvasFactory()
-  const canvasAndContext = canvasFactory.create(1, 1)
-
-  // Generate images
-  const images: Buffer[] = []
-  for (let idx = 1; idx <= numPages; idx += 1) {
-    const page = await pdfDocument.getPage(idx)
-    const viewport = getPageViewPort(page, opts.scaleImage)
-    canvasFactory.reset(canvasAndContext, viewport.width, viewport.height)
-    // TODO: fix types
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    await page.render({ canvasContext: canvasAndContext.context, viewport }).promise
-    page.cleanup()
-    const image = canvasAndContext.canvas.toBuffer('image/png')
-    images.push(image)
-  }
-
-  return Promise.all(images.map((x) => read(x)))
-}
-
-export const writeImages =
-  (outputImagePath: string, combinePages = true) =>
-  (images: ReadonlyArray<Jimp>): Promise<ReadonlyArray<Jimp>> => {
-    if (combinePages === true) {
-      return mergeImages(images)
-        .writeAsync(outputImagePath)
-        .then(() => images)
-    }
-
-    const parsedPath = path.parse(outputImagePath)
-    const partialName = path.join(parsedPath.dir, parsedPath.name)
-    const padMaxLen = images.length.toString().length
-    return Promise.all(
-      images.map((img, idx) =>
-        img.writeAsync(`${partialName}_${String(idx + 1).padStart(padMaxLen, '0')}.png`),
-      ),
-    ).then(() => images)
-  }
