+++ src/pdf2png/index.ts
@@ -1,0 +1,1 @@
+export { pdf2png } from './pdf2png'
